refactor: 整理 DB 連線與錯誤處理引用函式
